+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### v1.1.0 Feb 27, 2019
+
+* Support Square In-App Payments iOS and Android SDK 1.1.0.
+* For SDK 1.1.0 change, please check this [Change Log](https://docs.connect.squareup.com/changelog/mobile-logs/2019-02-27).
+
 ### v1.0.3 Feb 14, 2019
 
 * Compatible to AndroidX migration.